- Refactored Notion client by moving static functions to a separate file
